verzweiflung

+++ modulus/distributed/utils_mpi.py
@@ -291,7 +291,7 @@ def all_gather_v_wrapper(
     # Reconstruct the global tensor
     # Assuming the tensor is 1D for simplicity. Adjust for actual dimensions.
     global_tensor = torch.from_numpy(recv_buf).view(-1, *tensor.size()[1:]).to(tensor.device)
-
+    global_tensor=global_tensor.type(tensor.dtype)
     return global_tensor
     # tensor_shape = list(tensor.shape)
     # tensor_format = get_memory_format(tensor)
